requirements.txt
# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0
scipy>=1.10.0

# Machine Learning and Statistical Modeling
scikit-learn>=1.3.0
statsmodels>=0.14.0
lifetimes>=0.11.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter and Notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
ipython>=8.12.0

# Code Quality and Testing
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
pylint>=2.17.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Data Download and Management
kaggle>=1.5.0
requests>=2.31.0
